CNN-LSTM/requirements.txt
matplotlib
nltk
torch
numpy
pickle
Pillow
argparse
torchvision
